requirements.txt
Flask==3.0.0
imgurpython==1.1.7
line_bot_sdk==3.5.1
opencv_python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
pywin32==306
